Support returning the full test list with limit=-1
